Replace Bintray implementation
